requirements.txt
# -------------------------------
# 📦 Core
# -------------------------------
streamlit>=1.37.0
pandas>=2.2.0
numpy>=1.26.0
matplotlib>=3.9.0

# -------------------------------
# ⚙️ 모델 & 분석
# -------------------------------
scikit-learn>=1.5.0      # utils.py 내 모델 로딩 및 예측용
joblib>=1.3.0            # 모델 파일 로드
lightgbm>=4.3.0          # (사용 중이면 유지)
xgboost>=2.0.0           # (옵션)

# -------------------------------
# 🧠 LLM / RAG 기능 (필요 시)
# -------------------------------
chromadb==0.5.5
sentence-transformers==3.0.1
transformers>=4.41.0
torch>=2.2.0
llama-cpp-python>=0.2.86
requests>=2.32.0
python-dotenv>=1.0.1

# -------------------------------
# 📊 시각화 / 통계 확장
# -------------------------------
seaborn>=0.13.0          # (선택: 그래프 스타일링)
plotly>=5.22.0           # (선택: 인터랙티브 차트용)

# -------------------------------
# 🧩 기타 유틸
# -------------------------------
pydantic>=2.7.0
tqdm>=4.66.0
